Climber fixes
Removed soft limits on climber as they were not working if the climber was in the wrong position on startup, plus the climber does not seem to have an issue on hitting those limits.
Changed driver interface as per Mim's request. left and right sticks will drive the separate climbers all the time, pressing X will then drive both climbers with the left stick.